Microservices/requirements.txt
Django>=5.2
djangorestframework>=3.14.0
mysqlclient>=2.2.0
requests>=2.31.0
django-cors-headers>=4.3.0
